+++ cmd/powd/main.go
@@ -211,7 +211,8 @@ func setupLogging(repoPath string) error {
 		"ffs-coreipfs",
 		"ffs-grpc-service",
 		"ffs-filcold",
-		"ffs-sched-jstore",
+		"ffs-sched-sjstore",
+		"ffs-sched-rjstore",
 		"ffs-sched-astore",
 		"ffs-cidlogger",
 	}
